# specs/api/submission-api.openapi.yaml
openapi: 3.1.0
info:
  title: Gurine Submission API
  version: 13.0.0
  description: token-scoped external submissions. This is the final semantic API contract;
    Rust-generated OpenAPI must be equivalent.
servers:
- url: http://submission-api:8082
  description: Internal-only BFF-to-Submission-API origin
...
paths:
...
  /v1/correction-receipt:
    get:
      operationId: getCorrectionReceipt
      summary: 정정 요청 영수증
      tags:
      - submission
      x-operation-kind: QUERY
      x-error-codes:
      - INVALID_PARAMETER
      - RATE_LIMITED
      - SERVICE_ASSERTION_REQUIRED
      - SERVICE_ASSERTION_INVALID
      - SERVICE_ASSERTION_EXPIRED
      - SERVICE_ASSERTION_AUDIENCE_MISMATCH
      - SERVICE_ASSERTION_REQUEST_MISMATCH
      - SERVICE_ASSERTION_REPLAYED
      - SUBMISSION_SESSION_REQUIRED
      - SUBMISSION_SESSION_INVALID
      - SUBMISSION_SESSION_EXPIRED
      - SUBMISSION_SESSION_REVOKED
      - SUBMISSION_SESSION_SCOPE_MISMATCH
      - SNAPSHOT_STALE
      - RESOURCE_NOT_FOUND
      - PUBLICATION_GATE_BLOCKED
      - INTERNAL_ERROR
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectionReceiptResponse'
        '400':
          description: Problem response
          x-error-codes:
          - INVALID_PARAMETER
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Problem response
          x-error-codes:
          - SERVICE_ASSERTION_REQUIRED
          - SERVICE_ASSERTION_INVALID
          - SERVICE_ASSERTION_EXPIRED
          - SERVICE_ASSERTION_AUDIENCE_MISMATCH
          - SERVICE_ASSERTION_REQUEST_MISMATCH
          - SUBMISSION_SESSION_REQUIRED
          - SUBMISSION_SESSION_INVALID
          - SUBMISSION_SESSION_EXPIRED
          - SUBMISSION_SESSION_REVOKED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Problem response
          x-error-codes:
          - SUBMISSION_SESSION_SCOPE_MISMATCH
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Problem response
          x-error-codes:
          - RESOURCE_NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Problem response
          x-error-codes:
          - SERVICE_ASSERTION_REPLAYED
          - SNAPSHOT_STALE
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Problem response
          x-error-codes:
          - PUBLICATION_GATE_BLOCKED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '429':
          description: Problem response
          x-error-codes:
          - RATE_LIMITED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Problem response
          x-error-codes:
          - INTERNAL_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - BffServiceAssertion: []
        ScopedSubmissionSession: []
...
  /v1/response-receipt:
    get:
      operationId: getResponseReceipt
      summary: 접수 영수증
      tags:
      - submission
      x-operation-kind: QUERY
      x-error-codes:
      - INVALID_PARAMETER
      - RATE_LIMITED
      - SERVICE_ASSERTION_REQUIRED
      - SERVICE_ASSERTION_INVALID
      - SERVICE_ASSERTION_EXPIRED
      - SERVICE_ASSERTION_AUDIENCE_MISMATCH
      - SERVICE_ASSERTION_REQUEST_MISMATCH
      - SERVICE_ASSERTION_REPLAYED
      - SUBMISSION_SESSION_REQUIRED
      - SUBMISSION_SESSION_INVALID
      - SUBMISSION_SESSION_EXPIRED
      - SUBMISSION_SESSION_REVOKED
      - SUBMISSION_SESSION_SCOPE_MISMATCH
      - RESOURCE_NOT_FOUND
      - INTERNAL_ERROR
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseReceiptResponse'
        '400':
          description: Problem response
          x-error-codes:
          - INVALID_PARAMETER
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Problem response
          x-error-codes:
          - SERVICE_ASSERTION_REQUIRED
          - SERVICE_ASSERTION_INVALID
          - SERVICE_ASSERTION_EXPIRED
          - SERVICE_ASSERTION_AUDIENCE_MISMATCH
          - SERVICE_ASSERTION_REQUEST_MISMATCH
          - SUBMISSION_SESSION_REQUIRED
          - SUBMISSION_SESSION_INVALID
          - SUBMISSION_SESSION_EXPIRED
          - SUBMISSION_SESSION_REVOKED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Problem response
          x-error-codes:
          - SUBMISSION_SESSION_SCOPE_MISMATCH
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Problem response
          x-error-codes:
          - RESOURCE_NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Problem response
          x-error-codes:
          - SERVICE_ASSERTION_REPLAYED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '429':
          description: Problem response
          x-error-codes:
          - RATE_LIMITED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Problem response
          x-error-codes:
          - INTERNAL_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - BffServiceAssertion: []
        ScopedSubmissionSession: []
...
  /v1/response-session/access-status:
    get:
      operationId: getResponseAccessStatus
      summary: safe unavailable reason
      tags:
      - submission
      x-operation-kind: QUERY
      x-error-codes:
      - INVALID_PARAMETER
      - RATE_LIMITED
      - SERVICE_ASSERTION_REQUIRED
      - SERVICE_ASSERTION_INVALID
      - SERVICE_ASSERTION_EXPIRED
      - SERVICE_ASSERTION_AUDIENCE_MISMATCH
      - SERVICE_ASSERTION_REQUEST_MISMATCH
      - SERVICE_ASSERTION_REPLAYED
      - SUBMISSION_SESSION_REQUIRED
      - SUBMISSION_SESSION_INVALID
      - SUBMISSION_SESSION_EXPIRED
      - SUBMISSION_SESSION_REVOKED
      - SUBMISSION_SESSION_SCOPE_MISMATCH
      - RESOURCE_NOT_FOUND
      - INTERNAL_ERROR
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseAccessStatusResponse'
        '400':
          description: Problem response
          x-error-codes:
          - INVALID_PARAMETER
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Problem response
          x-error-codes:
          - SERVICE_ASSERTION_REQUIRED
          - SERVICE_ASSERTION_INVALID
          - SERVICE_ASSERTION_EXPIRED
          - SERVICE_ASSERTION_AUDIENCE_MISMATCH
          - SERVICE_ASSERTION_REQUEST_MISMATCH
          - SUBMISSION_SESSION_REQUIRED
          - SUBMISSION_SESSION_INVALID
          - SUBMISSION_SESSION_EXPIRED
          - SUBMISSION_SESSION_REVOKED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Problem response
          x-error-codes:
          - SUBMISSION_SESSION_SCOPE_MISMATCH
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Problem response
          x-error-codes:
          - RESOURCE_NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Problem response
          x-error-codes:
          - SERVICE_ASSERTION_REPLAYED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '429':
          description: Problem response
          x-error-codes:
          - RATE_LIMITED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Problem response
          x-error-codes:
          - INTERNAL_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
      - BffServiceAssertion: []
        ScopedSubmissionSession: []
...
  /v1/subscription-session:unsubscribe:
    post:
      operationId: unsubscribe
      summary: 해지
      tags:
      - submission
      x-operation-kind: COMMAND
      x-error-codes:
      - INVALID_PARAMETER
      - INVALID_REQUEST
      - RATE_LIMITED
      - SERVICE_ASSERTION_REQUIRED
      - SERVICE_ASSERTION_INVALID
      - SERVICE_ASSERTION_EXPIRED
      - SERVICE_ASSERTION_AUDIENCE_MISMATCH
      - SERVICE_ASSERTION_REQUEST_MISMATCH
      - SERVICE_ASSERTION_REPLAYED
      - SUBMISSION_SESSION_REQUIRED
      - SUBMISSION_SESSION_INVALID
      - SUBMISSION_SESSION_EXPIRED
      - SUBMISSION_SESSION_REVOKED
      - SUBMISSION_SESSION_SCOPE_MISMATCH
      - IDEMPOTENCY_CONFLICT
      - RESOURCE_NOT_FOUND
      - PRECONDITION_FAILED
      - VALIDATION_FAILED
      - INTERNAL_ERROR
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        schema:
          type: string
          format: uuid
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 8
          maxLength: 200
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/unsubscribeReceipt'
        '400':
          description: Problem response
          x-error-codes:
          - INVALID_PARAMETER
          - INVALID_REQUEST
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Problem response
          x-error-codes:
          - SERVICE_ASSERTION_REQUIRED
          - SERVICE_ASSERTION_INVALID
          - SERVICE_ASSERTION_EXPIRED
          - SERVICE_ASSERTION_AUDIENCE_MISMATCH
          - SERVICE_ASSERTION_REQUEST_MISMATCH
          - SUBMISSION_SESSION_REQUIRED
          - SUBMISSION_SESSION_INVALID
          - SUBMISSION_SESSION_EXPIRED
          - SUBMISSION_SESSION_REVOKED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Problem response
          x-error-codes:
          - SUBMISSION_SESSION_SCOPE_MISMATCH
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Problem response
          x-error-codes:
          - RESOURCE_NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Problem response
          x-error-codes:
          - SERVICE_ASSERTION_REPLAYED
          - IDEMPOTENCY_CONFLICT
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Problem response
          x-error-codes:
          - PRECONDITION_FAILED
          - VALIDATION_FAILED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '429':
          description: Problem response
          x-error-codes:
          - RATE_LIMITED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Problem response
          x-error-codes:
          - INTERNAL_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/unsubscribeRequest'
      security:
      - BffServiceAssertion: []
        ScopedSubmissionSession: []
...
components:
  securitySchemes:
    BffServiceAssertion:
      type: apiKey
      in: header
      name: X-Gurine-Service-Assertion
      description: Single-use request-bound Gurine Service Assertion issued by public-web
        or response-portal for audience submission-api.
    ScopedSubmissionSession:
      type: apiKey
      in: header
      name: X-Gurine-Submission-Session
      description: Opaque scoped session token forwarded only by an authenticated
        SvelteKit BFF; browsers store it only inside an authenticated-encrypted host-only
        cookie.
  schemas:
...
    CorrectionReceiptResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        version:
          type: integer
          format: int64
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        title:
          type: string
        summary:
          type: string
        data:
          $ref: '#/components/schemas/CorrectionReceipt'
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
      required:
      - id
      - status
      - data
      - links
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required:
      - type
      - title
      - status
      - requestId
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          format: uuid
        code:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        retryAfterSeconds:
          type: integer
...
    ResponseReceiptResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        version:
          type: integer
          format: int64
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        title:
          type: string
        summary:
          type: string
        data:
          $ref: '#/components/schemas/ResponseReceipt'
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
      required:
      - id
      - status
      - data
      - links
...
    unsubscribeRequest:
      type: object
      additionalProperties: false
      properties: {}
...